Added ability to select bluetooth (bt) mode
Added ability to select bluetooth (bt) as a mode option for the
controller. This change was made to the getMode function. The bluetooth
mode was set to mode number 4 and the other mode numbers were adjusted
accordingly. The getModePref function was updated to add support for the
bluetooth mode (it simply returns a number indicating bt mode was
selected for now, we don't call a modePrefFunction) and to accommodate
new mode numbers.