Finished getBaseShutter function
Added ability for user to select the base shutter used in HDR over
analog connections. Function returns the shutter time in micro-seconds
using a long value instead of an integer. An integer is too short.
Function only runs if the user selects and HDR bracket greater than 1
(they want to do HDR) and the connection is analog. The conditionals are
tested in the main (loop) function, not the getBaseShutter function.